+++ client-kotlin/src/test/kotlin/com/influxdb/client/kotlin/ITQueryKotlinApi.kt
@@ -110,7 +110,6 @@ internal class ITQueryKotlinApi : AbstractITInfluxDBClientKotlin() {
 
         val writeApi = client.writeApi
         writeApi.writeRecord(bucket.name, organization.id, WritePrecision.NS, records)
-        writeApi.close()
 
         client.close()