+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 1. [#33](https://github.com/influxdata/influxdb-client-java/issues/33): InfluxDBClient.close also dispose a created writeApi
+1. [#80](https://github.com/influxdata/influxdb-client-java/issues/80): FluxRecord, FluxColumn, FluxTable are serializable
 
 ## 1.4.0 [2020-01-17]