+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 1. [#33](https://github.com/influxdata/influxdb-client-java/issues/33): InfluxDBClient.close also dispose a created writeApi
 1. [#80](https://github.com/influxdata/influxdb-client-java/issues/80): FluxRecord, FluxColumn, FluxTable are serializable
 
+### Bugs
+1. [#82](https://github.com/influxdata/influxdb-client-java/pull/82): Apply backpressure strategy when a buffer overflow
+
 ## 1.4.0 [2020-01-17]
 
 ### Features