+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.5.0 [unreleased]
+## 1.5.0 [2020-02-14]
 
 ### Features
 1. [#33](https://github.com/influxdata/influxdb-client-java/issues/33): InfluxDBClient.close also dispose a created writeApi